agent/requirements.txt
livekit-agents[openai,silero]>=0.12.0
livekit-plugins-openai>=0.10.0
livekit-plugins-simli>=0.1.0
livekit>=0.17.0
openai>=1.50.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
websockets>=12.0
numpy>=1.26.0
